langgraph-agent/requirements.txt
langgraph>=0.2.76
langchain-openai>=0.2.0
langchain-community>=0.3.0  
langchain-core>=0.3.0
tavily-python>=0.5.0
duckduckgo-search>=6.0.0
yfinance>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0